+++ src/main/java/org/eclipse/keyple/card/calypso/CmdCardSvReload.java
@@ -52,16 +52,12 @@ final class CmdCardSvReload extends AbstractCardCommand {
   /** The command. */
   private static final CalypsoCardCommand command = CalypsoCardCommand.SV_RELOAD;
 
-  private static final int SV_POSTPONED_DATA_IN_SESSION = 0x6200;
+  private static final int SW_POSTPONED_DATA = 0x6200;
   private static final Map<Integer, StatusProperties> STATUS_TABLE;
 
   static {
     Map<Integer, StatusProperties> m =
         new HashMap<Integer, StatusProperties>(AbstractApduCommand.STATUS_TABLE);
-    m.put(
-        SV_POSTPONED_DATA_IN_SESSION,
-        new StatusProperties(
-            "Successful execution, response data postponed until session closing.", null));
     m.put(
         0x6400,
         new StatusProperties(
@@ -79,6 +75,10 @@ final class CmdCardSvReload extends AbstractCardCommand {
         new StatusProperties(
             "Preconditions not satisfied.", CalypsoSamAccessForbiddenException.class));
     m.put(0x6988, new StatusProperties("Incorrect signatureHi.", CardSecurityDataException.class));
+    m.put(
+        SW_POSTPONED_DATA,
+        new StatusProperties(
+            "Successful execution, response data postponed until session closing."));
     STATUS_TABLE = m;
   }
 
@@ -186,7 +186,7 @@ final class CmdCardSvReload extends AbstractCardCommand {
                     p2,
                     dataIn,
                     null))
-            .addSuccessfulStatusWord(SV_POSTPONED_DATA_IN_SESSION));
+            .addSuccessfulStatusWord(SW_POSTPONED_DATA));
   }
 
   /**